fix: satisfy backend-size limits for cli server flag consolidation
Split run --server bind tests into root_run_server_test.go and extract
submit payload preparation so CI backend-size verification passes.
